requirements.txt
spotipy~=2.23.0
Gooey~=1.0.8.1
sortedcontainers~=2.4.0
python-dotenv~=1.0.1
requests_html==0.10.0
yt_dlp~=2024.04.09
lxml[html_clean]
requests~=2.31.0